+++ Sources/QueryParser.swift
@@ -164,10 +164,7 @@ public class QueryParser {
         var sq = ""
         while (!tq.isEmpty()) {
             if tq.matchesCS("\\") {
-                sq.append(tq.consume())
-                if !tq.isEmpty() {
-                    sq.append(tq.consume())
-                }
+                sq.append(tq.consumeCssEscapeSequence())
             } else if (tq.matches("(")) {
                 sq.append("(")
                 sq.append(tq.chompBalanced("(", ")"))